requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.4
django-cors-headers==4.3.1
python-dotenv==1.0.1
gunicorn==22.0.0 